FIx linters
