+++ src/index.js
@@ -9,10 +9,10 @@ const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <BrowserRouter>
-    <Provider store={store}>
+      <Provider store={store}>
         <App />
-      </Provider>   
+      </Provider>
     </BrowserRouter>
- 
+
   </React.StrictMode>,
 );
