+++ src/redux/store.js
@@ -1,10 +1,10 @@
-import { configureStore } from "@reduxjs/toolkit";
-import greetingReducer from "./features/greetingsSlice";
+import { configureStore } from '@reduxjs/toolkit';
+import greetingReducer from './features/greetingsSlice';
 
 const store = configureStore({
-    reducer : {
-        greeting: greetingReducer,
-    },
+  reducer: {
+    greeting: greetingReducer,
+  },
 });
 
 export default store;
